Added mongo integration and removed elastic search

+++ indexservice/src/main/java/com/searchengine/indexservice/models/HtmlDocument.java
@@ -15,7 +15,7 @@ import java.util.List;
 @AllArgsConstructor
 @Builder
 public class HtmlDocument {
-    private Integer urlId;
+    private Long urlId;
     private String url;
     private String title;
     private String body;
