+++ indexservice/src/main/java/com/searchengine/indexservice/models/SQSHtmlMetadata.java
@@ -12,7 +12,7 @@ import org.springframework.http.HttpStatus;
 @NoArgsConstructor
 @AllArgsConstructor
 public class SQSHtmlMetadata {
-    private Integer urlId;
+    private Long urlId;
     private String url;
     private String redirectedUrl;
     private HttpStatus status;
